refactor(qmc): extract qm meta code to utils
